ml-models/yolo/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
pyyaml==6.0.1
scipy==1.12.0
matplotlib==3.8.2